requirements.txt
llama-cpp-python
lark
pydantic
